refactored upload logic, external packages, and services for uploading pictures on /upload/image route
